Ensure code works even without typing_extensions installed
